+++ insanic/scopes.py
@@ -1,7 +1,11 @@
 import os
+import urllib.request
 
 from functools import wraps
 
+from insanic.log import logger
+
+AWS_ECS_METADATA_ENDPOINT = "169.254.170.2/v2/metadata"
 
 def public_facing(f):
     @wraps(f)
@@ -13,14 +17,18 @@ def public_facing(f):
 
 def _is_docker():
     try:
-        with open('/proc/self/cgroup', 'r') as proc_file:
-            for line in proc_file:
-                fields = line.strip().split('/')
-                if fields[1] == 'docker':
-                    return True
-    except FileNotFoundError:
-        pass
-
+        r = urllib.request.urlopen("http://" + AWS_ECS_METADATA_ENDPOINT, timeout=0.5)
+        logger.info(r.read().decode())
+        return r.status == 200
+    except:
+        try:
+            with open('/proc/self/cgroup', 'r') as proc_file:
+                for line in proc_file:
+                    fields = line.strip().split('/')
+                    if fields[1] == 'docker':
+                        return True
+        except FileNotFoundError:
+            pass
     return False